+++ docs/history.rst
@@ -7,6 +7,9 @@ Latest
 - REF: Handle deprecation of proj_context_set_autoclose_database (issue #866)
 - DOC: Improve FAQ text about CRS formats (issue #789)
 - BUG: Add PyPy cython array implementation (issue #854)
+- BUG: Fix spelling for
+  :class:`pyproj.crs.coordinate_operation.AzimuthalEquidistantConversion`
+  and :class:`pyproj.crs.coordinate_operation.LambertAzimuthalEqualAreaConversion` (issue #882)
 
 3.1.0
 -----